Refactored DevotionDisplay by moving date conversion and month sorting logic to new devotionUtils.ts for better readability and maintainability.
